Update with latest features 1/3
This is a global update to add multiple experience to explore the implementation details. See more details in the updated README.

This includes update to use latest Meta XR SDKs v68
Fix all issues related to the Unity 2022 updates

Added use cases for VoiceSDK, haptics, locomotion, and passthrough.

Reviewed By: kpfeifer-meta

Differential Revision: D64214504

fbshipit-source-id: 63191cbcb190b159684626d0f562e145756fb798

+++ Assets/Project/Scripts/UI/UICanvas.cs
@@ -1,26 +1,9 @@
-﻿/*
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- * All rights reserved.
- *
- * Licensed under the Oculus SDK License Agreement (the "License");
- * you may not use the Oculus SDK except in compliance with the License,
- * which is provided at the time of installation or download, or which
- * otherwise accompanies this software in either electronic or hard copy form.
- *
- * You may obtain a copy of the License at
- *
- * https://developer.oculus.com/licenses/oculussdk/
- *
- * Unless required by applicable law or agreed to in writing, the Oculus SDK
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright (c) Meta Platforms, Inc. and affiliates.
 
 using System;
 using System.Threading.Tasks;
 using UnityEngine;
+using UnityEngine.UI;
 
 namespace Oculus.Interaction.ComprehensiveSample
 {
@@ -29,21 +12,41 @@ namespace Oculus.Interaction.ComprehensiveSample
     /// via Show and ShowAsync methods, used to show or hide UI
     /// An IUICanvasAnimator component can be added to override the default fade animation
     /// </summary>
-    [RequireComponent(typeof(Canvas)), RequireComponent(typeof(CanvasGroup))]
-    public sealed class UICanvas : MonoBehaviour, IActiveState
+    [RequireComponent(typeof(CanvasGroup))]
+    public class UICanvas : MonoBehaviour, IActiveState
     {
         [SerializeField]
         private bool _show = true;
         [SerializeField]
         private float _duration = 0.3f;
+        [SerializeField, Tooltip("Controls if the canvas enabled/gameobject active state should be controlled by this")]
+        private DisableMode _disableMethod = DisableMode.Canvas;
 
-        private Canvas _canvas;
-        private CanvasGroup _canvasGroup;
-        private IUICanvasAnimator _canvasAnimator;
-        private TaskCompletionSource<bool> _canvasAnimatorTask;
+        Canvas _canvas;
+        CanvasGroup _canvasGroup;
+        bool _hasGraphicRaycaster;
+        GraphicRaycaster _graphicRaycaster;
+        IUICanvasAnimator _canvasAnimator;
+        TaskCompletionSource<bool> _canvasAnimatorTask;
         private bool _started;
 
-        public bool Active => _show;
+        bool IActiveState.Active => _show;
+
+        public event Action onChange;
+
+        public float Duration
+        {
+            get => _duration;
+            set => _duration = value;
+        }
+
+        public bool IsShown
+        {
+            get => _show;
+            set => Show(value);
+        }
+
+        private void Reset() => RequireCanvas();
 
         private void Awake()
         {
@@ -52,6 +55,11 @@ namespace Oculus.Interaction.ComprehensiveSample
             _duration = 0;
             Show(!(_show = !_show));
             _duration = duration;
+        }
+
+        void OnDestroy()
+        {
+            TweenRunner.Kill(_canvasGroup);
         }
 
         public async void Show(bool value)
@@ -65,6 +73,12 @@ namespace Oculus.Interaction.ComprehensiveSample
 
             _show = show;
 
+            if (_started) onChange?.Invoke();
+
+            //Task has completed
+            if (Application.isPlaying == false)
+                return true;
+
             EnsureReferences();
             TweenRunner.Kill(_canvasGroup);
 
@@ -77,8 +91,11 @@ namespace Oculus.Interaction.ComprehensiveSample
                 task.TrySetResult(false);
             }
 
-            if (show) _canvas.enabled = true;
+            if (show && (_disableMethod & DisableMode.Canvas) != 0) _canvas.enabled = true;
+            if (show && (_disableMethod & DisableMode.GameObject) != 0) _canvas.gameObject.SetActive(true);
+
             _canvasGroup.blocksRaycasts = false;
+            if (_hasGraphicRaycaster) _graphicRaycaster.enabled = false;
 
             bool finished = true;
 
@@ -102,6 +119,7 @@ namespace Oculus.Interaction.ComprehensiveSample
                 if (duration > 0)
                 {
                     finished = await TweenRunner.Tween(_canvasGroup.alpha, show ? 1f : 0f, duration, x => _canvasGroup.alpha = x)
+                        .IgnoreTimeScale()
                         .SetID(_canvasGroup)
                         .ToTask();
                 }
@@ -110,7 +128,9 @@ namespace Oculus.Interaction.ComprehensiveSample
             if (finished)
             {
                 if (!hasAnimator) _canvasGroup.alpha = show ? 1f : 0f;
-                if (!show) _canvas.enabled = false;
+                if (!show && (_disableMethod & DisableMode.Canvas) != 0) _canvas.enabled = false;
+                if (!show && (_disableMethod & DisableMode.GameObject) != 0) _canvas.gameObject.SetActive(false);
+                if (_hasGraphicRaycaster) _graphicRaycaster.enabled = show;
                 _canvasGroup.blocksRaycasts = show;
             }
 
@@ -125,6 +145,34 @@ namespace Oculus.Interaction.ComprehensiveSample
             _canvas = GetComponent<Canvas>();
             _canvasGroup = GetComponent<CanvasGroup>();
             _canvasAnimator = GetComponent<IUICanvasAnimator>();
+            _hasGraphicRaycaster = TryGetComponent(out _graphicRaycaster);
+        }
+
+        /// <summary>
+        /// An alternative to RequireComponent(typeof(Canvas)) that will add a a canvas component but also add a GraphicRaycaster if needed
+        /// </summary>
+        private void RequireCanvas()
+        {
+            if (TryGetComponent<Canvas>(out var _)) { return; }
+
+            var parentCanvas = GetComponentInParent<Canvas>();
+            var parentRaycaster = parentCanvas ? parentCanvas.GetComponent<GraphicRaycaster>() : null;
+
+            gameObject.AddComponent<Canvas>();
+            if (parentRaycaster)
+            {
+                var newRaycaster = gameObject.AddComponent<GraphicRaycaster>();
+                newRaycaster.blockingMask = parentRaycaster.blockingMask;
+                newRaycaster.blockingObjects = parentRaycaster.blockingObjects;
+                newRaycaster.ignoreReversedGraphics = parentRaycaster.ignoreReversedGraphics;
+            }
+        }
+
+        [Flags]
+        public enum DisableMode
+        {
+            Canvas = 1 << 0,
+            GameObject = 1 << 1,
         }
     }
